Add better CSS support
This was a more complicated portion of the tutorial. A lot of
things needed to be changed and added to get the desired
functionality.  Stylesheets can now be added through importing and
can be used by calling their className.

A lot of changes were made to webpack.config.js to get the
necessary functionality.