Create functionality for multiple environments
Because we will be interfacing with the Google API, we need to
make sure we don't hardcode our keys into a deployed application.
This commit lays the groundwork for us to configure our app to
include dynamic API keys based on the environment.  The changes
contained in webpack should allow our app to be able to access
environment variables so we can store our API keys within our
computers envornment variables rather than configuration files
which will be uploaded to github.