Add groundwork for testing framework
Added a set of testing modules to help with TDD.  Most of the work
is done through Karma.  A lot of changes and additions had to be
made in order to get this to work, mostly in webpack and the karma
configuration files.  Some of the app code needed to be moved to
the containers folder.